Assignment3_minor_fix
add template requires template argument list

+++ Assignment3/QueueStack.h
@@ -161,7 +161,7 @@ namespace assignment3
 	double QueueStack<T>::GetAverage() const
 	{
 		assert(mSmartStacks.size() > 0 && mSmartStacks.front().GetCount() > 0);
-		return GetSum() / GetCount();
+		return GetSum() / static_cast<double>(GetCount());
 	}
 
 	template <typename T>
